Login
Changed Spoors container to fetch the viewer from server instead.
If there is no `viewer.user`, the avatar displays a generic icon, and
clicking on it will display the login form.

The form will perform a mutation upon submit (through a prop in the
Login container), which asks the server to create a JWT token if the
credentials are correct. This token will be set to `AsyncStorage`.

When the app loads, there is now a middle ware in Apollo network layer
to check for existing token in `AsyncStorage` and set `authorization`
header to that token.

The server should check that header and verify the token and assign
the user (with at least the `id` property) to the GraphQL root value.
This root value is used to return the user in `viewer.user` or not.